Changes to make things compile with MSVC
 * Add #include <array> in lots of places, apparently libstdc++
   brings it in transitively
 * Remove constraints from lens_base, MSVC doesn't seem to like it
 * Remove the neat hack to use string_view as its own subsequence type,
   as MSVC specifically checks for it and disallows it
 * Use C++23 mode in CMakeLists.txt if the compiler is MSVC, because
   ranges isn't ABI stable yet

Overall though, vastly fewer changes than I thought would be required,
and no constexpr workarounds needed -- MSVC has come a long long way
in the last couple of years.

+++ test/test_bitset.cpp
@@ -10,6 +10,7 @@
 
 #include "test_utils.hpp"
 
+#include <array>
 #include <iostream>
 
 namespace {
@@ -36,7 +37,7 @@ constexpr bool test_bitset()
         uint32_t x = 0;
         FLUX_FOR(bool bit, flux::reverse(std::as_const(b))) {
             x <<= 1;
-            x |= bit;
+            x |= uint32_t(bit);
         }
 
         STATIC_CHECK(x == 0xCAFEBABE);
